# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Payment gateway aggregator — VA, QRIS, and disbursement via Midtrans, Xendit, DOKU, iPaymu.\n\n## Authentication\nMerchant endpoints (/v1/*): `X-API-Key` header with your API key (`wpay_live_*` or `wpay_test_*`).\nAdmin endpoints (/admin/*): `Authorization: Bearer \u003caccess_token\u003e` from POST /admin/login.\n\n## Idempotency\nPayment and disbursement creation support `Idempotency-Key` header. Use unique key per request to safely retry.\n\n## Webhook\nWanpey signs outbound webhooks with `X-Wanpey-Signature` (HMAC-SHA256). Verify using your webhook secret.",
        "title": "Wanpey Payment Gateway API",
        "contact": {
            "name": "Wanpey Support",
            "email": "support@wanpey.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/mutations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns paginated ledger mutations (credits and debits) for the authenticated merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mutations (Ledger)"
                ],
                "summary": "List ledger mutations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type: credit, debit",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wanpey_core_internal_usecase.MutationOutput"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/mutations/balance": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the merchant's current spendable balance in IDR (calculated live from ledger).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mutations (Ledger)"
                ],
                "summary": "Get current balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "balance": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/mutations/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a single ledger mutation by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mutations (Ledger)"
                ],
                "summary": "Get mutation detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mutation ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.MutationOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wanpey_core_internal_domain_entity.Currency": {
            "type": "string",
            "enum": [
                "IDR"
            ],
            "x-enum-varnames": [
                "CurrencyIDR"
            ]
        },
...
        "wanpey_core_internal_domain_entity.MutationReferenceType": {
            "type": "string",
            "enum": [
                "payment",
                "disbursement"
            ],
            "x-enum-varnames": [
                "MutationRefPayment",
                "MutationRefDisbursement"
            ]
        },
        "wanpey_core_internal_domain_entity.MutationType": {
            "type": "string",
            "enum": [
                "cash_in",
                "cash_out"
            ],
            "x-enum-varnames": [
                "MutationTypeCashIn",
                "MutationTypeCashOut"
            ]
        },
...
        "wanpey_core_internal_usecase.MutationOutput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.Currency"
                },
                "description": {
                    "type": "string"
                },
                "fee_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "net_amount": {
                    "type": "integer"
                },
                "reference_id": {
                    "type": "string"
                },
                "reference_type": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MutationReferenceType"
                },
                "type": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MutationType"
                }
            }
        },
...
        "wanpey_core_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wanpey_core_pkg_response.FieldDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "wanpey_core_pkg_response.FieldDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "wanpey_core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Pagination"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "wanpey_core_pkg_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT — Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Merchant API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}